// apps/zhurong-platform-custom-but/src/api/order/typing.ts
export interface VimOrderlVO {
  belposId: string;

  itemCode: string;

  coutKey: string;

  cfccad: string;

  cfType: string;

  ccad: string;

  itemName: string;

  uZnr: string;

  udf1: number;

  aplatzId: string;

  mengeVerbrauch: number;

  anfzeit: string;

  lieferdatum: string;

  vversion: string;

  icadproduct: string;
}